backend/requirements.txt
asgiref==3.8.1
Brotli==1.1.0
click==8.1.7
dj-database-url==2.1.0
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
h11==0.14.0
jsonfield==3.1.0
packaging==24.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
uvicorn==0.29.0
whitenoise==6.6.0
